refactor: 회원가입 페이지 단일책임 원칙 적용/ 액션들 더 분리해도 될듯
